+++ blekit/src/main/java/com/outlook/wn123o/blekit/util/BleKitUtils.java
@@ -7,7 +7,7 @@ public final class BleKitUtils {
         throw new RuntimeException();
     }
 
-    public static UUID ble16BitUuid(long d) {
-        return new UUID(0x1000L |  (d << 32), 0x800000805f9B34FBL);
+    public static UUID ble16BitUuid(int d) {
+        return new UUID(0x1000L |  ((long) d << 32), 0x800000805f9B34FBL);
     }
 }
